Renamed.
Room generation works correctly now. It also appears to be fairly efficient in that it can handle the max value of attempts, with the minimum room size (1) and maximum number of tiles (screen width) quickly. The current limit of that isn't the algorithm itself, but the addition of tiles based on the rooms generated. However, that will be made more efficient soon.